requirements.txt
streamlit 
requests 
beautifulsoup4 
pandas 
transformers 
sentence-transformers 
faiss-cpu 
torch 
numpy